requirements.txt
fastapi==0.115.1
uvicorn==0.24.0.post1

requests==2.32.3
beautifulsoup4==4.12.3

pandas==2.2.2
pdfplumber==0.11.6

mysql-connector-python==8.3.0
dotenv
